requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.9
django-cors-headers==3.4.0
django-jsonfield==1.4.0
djangorestframework==3.11.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
